refactor: ReadHoldingRegisters is Working! :)

+++ ModbusNet/Transport/IModbusTransport.cs
@@ -4,7 +4,7 @@ namespace ModbusNet.Transport
 {
     public interface IModbusTransport : IDisposable
     {
-        byte[] SendRequest(byte[] request);
+        ushort[] SendRequest(byte[] request);
         bool IsConnected { get; }
         void Connect();
         void Disconnect();
